perform some renames
