+++ dashboard/src/components/Config/PkgRepoList/PkgRepoControl.tsx
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { CdsButton } from "@cds/react/button";
+import ConfirmDialog from "components/ConfirmDialog";
 import {
   PackageRepositoryReference,
   PackageRepositorySummary,
@@ -12,8 +13,7 @@ import { Action } from "redux";
 import { ThunkDispatch } from "redux-thunk";
 import { IStoreState } from "shared/types";
 import actions from "../../../actions";
-import ConfirmDialog from "../../ConfirmDialog/ConfirmDialog";
-import { PkgRepoAddButton } from "./PkgRepoButton";
+import { PkgRepoAddButton } from "./PkgRepoAddButton";
 import "./PkgRepoControl.css";
 
 export interface IPkgRepoListItemProps {
