+++ dashboard/src/components/Config/PkgRepoList/PkgRepoList.tsx
@@ -9,7 +9,7 @@ import { filterNames, filtersToQuery } from "components/Catalog/Catalog";
 import Alert from "components/js/Alert";
 import Table from "components/js/Table";
 import Tooltip from "components/js/Tooltip";
-import PageHeader from "components/PageHeader/PageHeader";
+import PageHeader from "components/PageHeader";
 import { push } from "connected-react-router";
 import { PackageRepositorySummary } from "gen/kubeappsapis/core/packages/v1alpha1/repositories";
 import qs from "qs";
@@ -21,7 +21,7 @@ import { IStoreState } from "shared/types";
 import { app } from "shared/url";
 import { getPluginName } from "shared/utils";
 import LoadingWrapper from "../../LoadingWrapper/LoadingWrapper";
-import { PkgRepoAddButton } from "./PkgRepoButton";
+import { PkgRepoAddButton } from "./PkgRepoAddButton";
 import { PkgRepoControl } from "./PkgRepoControl";
 import { PkgRepoDisabledControl } from "./PkgRepoDisabledControl";
 import "./PkgRepoList.css";
